make share button more generic
